finish file restructure
