+++ src/lib/registry/registry.tsx
